Introduce the various integer types from the zng spec
As of this commit, test coverage on the new types is thin, that will
come in a subsequent commit.

+++ zql/zql.js
@@ -241,7 +241,7 @@ function peg$parse(input, options) {
             return makeLiteral("float64", v)
           },
       peg$c40 = function(v) {
-            return makeLiteral("int", v)
+            return makeLiteral("int64", v)
           },
       peg$c41 = function(v) { return v },
       peg$c42 = function(v) {
